点击视频后下方的相关推荐视频列表

+++ app/src/main/java/com/example/openeyes/model/CommentModel.kt
@@ -30,7 +30,7 @@ data class CommentModel(
             val id: Long?,
             val videoId: Int?,
             val videoTitle: String?,
-            val parentReplyId: Int?,
+            val parentReplyId: Long?,
             val rootReplyId: Long?,
             val sequence: Int?,
             val message: String?,
